backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg==3.1.13
psycopg-binary==3.3.2
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client for OAuth
httpx==0.25.1

# GitHub Integration
PyGithub==2.1.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
pyod==1.1.2

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
